+++ HandsClothes/EFData/Unit.cs
@@ -20,7 +20,7 @@ namespace HandsClothes.EFData
             this.Material = new HashSet<Material>();
         }
     
-        public byte id { get; set; }
+        public int id { get; set; }
         public string UnitName { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
